requirements.txt
# Core ML framework
torch>=2.0.0
torchaudio>=2.0.0
pytorch-lightning>=2.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
webrtcvad>=2.0.10

# Data processing and augmentation
lhotse>=1.20.0
torch-audiomentations>=0.16.0
pyannote.audio>=2.1.0

# Utilities
numpy>=1.21.0
scipy>=1.9.0
matplotlib>=3.5.0
tqdm>=4.64.0
pyyaml>=6.0
hydra-core>=1.3.0
wandb>=0.15.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional: For neural codec features
# encodec>=0.1.0  # Uncomment if using EnCodec features 